[Remove] StepProgress 컴포넌트 통합 (StepSegment 제거)
